move functions into render.ts
